+++ odoo/songs/install/data_demo.py
@@ -118,6 +118,17 @@ def import_sales_order(ctx):
 
 
 @anthem.log
+def import_purchase_order(ctx):
+    """ Importing purchases order from csv """
+    model = ctx.env['purchase.order'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/purchase_order_head.csv', model)
+    load_csv(ctx, 'data/demo/purchase_order_line.csv',
+             'purchase.order.line')
+
+
+@anthem.log
 def main(ctx):
     """ Loading demo data """
     import_users(ctx)
@@ -137,4 +148,5 @@ def main(ctx):
     import_bank(ctx)
     import_bank_account(ctx)
     import_sales_order(ctx)
+    import_purchase_order(ctx)
     return